Moved mapping tests to functor
